consumerir: Dynamically grow an appropriately sized buffer
The IR pulse sequences are written to a local buffer to be
passed to the kernel.

Previously, we doubled the fixed size buffer (from 1K to 2K)
to work around a crash that was occurring for long sequences.

Another crash has been found when using Smart IR Remote and the
POWER function of the TX P42X10E remote by Panasonic (TV category).

Rather than double the size again and wait for another counter-
example, dynamically allocate the buffer and grow it as needed
to handle any IR sequence.

Change-Id: I314a23fc04f7c9dce402f04e469a7569fe4b7d05